Episodes code cleanup

+++ hikka/services/files.py
@@ -16,9 +16,10 @@ class FileService:
 
     @classmethod
     def destroy(cls, file: File):
-        if not file.is_link():
-            if file.uploaded and file.path:
-                fs = storage.init_fs()
-                fs.rm(file.storage())
+        if file:
+            if not file.is_link():
+                if file.uploaded and file.path:
+                    fs = storage.init_fs()
+                    fs.rm(file.storage())
 
-            file.delete()
+                file.delete()
